Final commit

+++ ByteCoin/Model/CoinData.swift
@@ -14,6 +14,6 @@ struct CoinData: Decodable {
     let time: String
     let asset_id_base: String
     let asset_id_quote: String
-    let rate: Float
+    let rate: Double
     
 } //struct CoinData
